kubernetes/apps/kfserving/examples/mobilenet/transformer/requirements.txt
kfserving>=0.2.1
argparse>=1.4.0
requests>=2.22.0
joblib>=0.13.2
pandas>=0.24.2
numpy>=1.16.3
kubernetes >= 9.0.0
pillow==6.2.0

tensorflow == 1.15.2
